Reading the JSON response from Media Wiki API
Implementation of a Java method that looks up the number of contributions done by a particular user in Wikidata and prints the result in the console.

Java program request the user to type the name of user whose contribution is need to be find. Then it scans the input from user and read the JSON response from Media Wiki API . The details of the contribution of the particular user is stored in the JSON file locally (“output.json”). Then Count method finally prints the number of contribution made by the user.